feat: add logout component to patient and doctor profile
