+++ src/components/LogOutBanner/LogOutBanner.module.scss
@@ -1,0 +1,14 @@
+.caption {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 10px;
+  border-radius: 5px;
+  padding: 15px;
+  height: 75px;
+  background-color: rgb(53 171 215);
+}
+
+.text {
+  font-weight: 100;
+}
