+++ src/pages/DoctorAccount/DoctorAccount.tsx
@@ -15,6 +15,7 @@ import MedicalInformation from "@mui/icons-material/MedicalInformation";
 import LocationCity from "@mui/icons-material/LocationCity";
 import CircularProgress from "@mui/material/CircularProgress";
 import style from "./DoctorAccount.module.scss";
+import { LogOutBanner } from "../../components/LogOutBanner/LogOutBanner";
 
 const openModal = () => {
   console.log("");
@@ -35,19 +36,23 @@ export const DoctorAccount = () => {
   );
 
   const { data: doctor, isLoading: isLoadingDoctor } =
-    useQuery<IDoctor, Error>(["doctor", id], () => getDoctor(id), {onError: () => {
-      setAlert({ severity: "error", message: "Error during fetching data" });
-    }});
+    useQuery<IDoctor, Error>(["doctor", id], () => getDoctor(id), {
+      onError: () => {
+        setAlert({ severity: "error", message: "Error during fetching data" });
+      }
+    });
 
   const onUploadAvatar = async (file: File) => {
     mutation.mutate(file);
   };
 
   const { data: infoAppointments, isLoading: isLoadingSchedule, isSuccess: isSuccessSchedule } =
-    useQuery<ISchedule, Error>(["schedule", id], () => getSchedule(id), {onError: () => {
-      setAlert({ severity: "error", message: "Error during fetching data" });
-    }});
-    console.log(infoAppointments?.schedule);
+    useQuery<ISchedule, Error>(["schedule", id], () => getSchedule(id), {
+      onError: () => {
+        setAlert({ severity: "error", message: "Error during fetching data" });
+      }
+    });
+  console.log(infoAppointments?.schedule);
 
   const [alert, setAlert] = useState<AlertType | null>(null);
   const [isModalOpen, setModalOpen] = useState(false);
@@ -58,61 +63,65 @@ export const DoctorAccount = () => {
   };
   const handleModalClose = () => setModalOpen(false);
 
+
   return (
-    <Box className={style.wrapper}>
-      <Box sx={{display: "flex", alignItems: "center"}} mb={3}>
-        <PhotoWithUpload isLoading={isLoadingDoctor} onUpload={onUploadAvatar} image={doctor?.photo} />
-        <Typography variant="h3" color={"primary"}>
-          {doctor?.nameDoctor} {doctor?.surname}
-        </Typography>
-      </Box>
-      <Button id={"edit"} onClick={handleModalOpen}>Edit</Button>
-      <Box className={style.infoBlock} mb={4} mt={1}>
-       <DescriptionField icon={MedicalInformation} caption={"Category:"} text={doctor?.category}/>
-       <DescriptionField icon={LocationCity} caption={"City:"} text={doctor?.city}/>
-      </Box>
-      <Grid container gap={2}>
-        <Grid className={style.infoBlock} item xs>
-          <Typography>About me:</Typography>
+    <>
+      <LogOutBanner />
+      <Box className={style.wrapper}>
+        <Box sx={{ display: "flex", alignItems: "center" }} mb={3}>
+          <PhotoWithUpload isLoading={isLoadingDoctor} onUpload={onUploadAvatar} image={doctor?.photo} />
+          <Typography variant="h3" color={"primary"}>
+            {doctor?.nameDoctor} {doctor?.surname}
+          </Typography>
+        </Box>
+        <Button id={"edit"} onClick={handleModalOpen}>Edit</Button>
+        <Box className={style.infoBlock} mb={4} mt={1}>
+          <DescriptionField icon={MedicalInformation} caption={"Category:"} text={doctor?.category} />
+          <DescriptionField icon={LocationCity} caption={"City:"} text={doctor?.city} />
+        </Box>
+        <Grid container gap={2}>
+          <Grid className={style.infoBlock} item xs>
+            <Typography>About me:</Typography>
+          </Grid>
+          <Grid className={style.infoBlock} item xs>
+            <Button id={"time"} onClick={handleModalOpen}>Сhoice of appointment time</Button>
+            {isLoadingSchedule ? (
+              <Box sx={{ display: "flex", justifyContent: "center", alignItems: "center", height: "200px" }}>
+                <CircularProgress />
+              </Box>
+            ) : (
+              <>
+                {isSuccessSchedule && infoAppointments ? (
+                  <SectionSchedule data={infoAppointments.schedule} onClick={openModal} />
+                ) : (<Typography mb={2}>{"No saved appointment time"}</Typography>)
+                }
+              </>
+            )}
+          </Grid>
         </Grid>
-        <Grid className={style.infoBlock} item xs>
-          <Button id={"time"} onClick={handleModalOpen}>Сhoice of appointment time</Button>
-          {isLoadingSchedule ? (
-            <Box sx={{ display: "flex", justifyContent: "center", alignItems: "center", height: "200px" }}>
-              <CircularProgress />
-            </Box>
-          ) : (
-            <>
-              {isSuccessSchedule && infoAppointments ? (
-                <SectionSchedule data={infoAppointments.schedule} onClick={openModal}/>
-              ) : (<Typography mb={2}>{"No saved appointment time"}</Typography>)
-              }
-            </>
-          )}
-        </Grid>
-      </Grid>
-      <Modal
-        open={isModalOpen}
-        onClose={handleModalClose}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description"
-      >
-        {
-          typeModal === "edit"
-          ? <EditDoctorModal data={doctor} id={id} onClose={handleModalClose} setAlert={setAlert} />
-          : typeModal === "time" && isSuccessSchedule
-            ? <EditDataModal data={infoAppointments.schedule}/>
-            : <></>
-        }
-      </Modal>
-      <Snackbar
-        open={!!alert}
-        autoHideDuration={3000}
-        onClose={() => setAlert(null)}
-        anchorOrigin={{ vertical: "bottom", horizontal: "center" }}
-      >
-        <Alert onClose={() => setAlert(null)} severity={alert?.severity || "success"}>{alert?.message}</Alert>
-      </Snackbar>
-    </Box>
+        <Modal
+          open={isModalOpen}
+          onClose={handleModalClose}
+          aria-labelledby="modal-modal-title"
+          aria-describedby="modal-modal-description"
+        >
+          {
+            typeModal === "edit"
+              ? <EditDoctorModal data={doctor} id={id} onClose={handleModalClose} setAlert={setAlert} />
+              : typeModal === "time" && isSuccessSchedule
+                ? <EditDataModal data={infoAppointments.schedule} />
+                : <></>
+          }
+        </Modal>
+        <Snackbar
+          open={!!alert}
+          autoHideDuration={3000}
+          onClose={() => setAlert(null)}
+          anchorOrigin={{ vertical: "bottom", horizontal: "center" }}
+        >
+          <Alert onClose={() => setAlert(null)} severity={alert?.severity || "success"}>{alert?.message}</Alert>
+        </Snackbar>
+      </Box>
+    </>
   );
 };
