+++ src/pages/PatientAccount/PatientAccount.module.scss
@@ -1,16 +1,3 @@
-.caption {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 15px;
-  height: 75px;
-  background-color: rgb(53 171 215);
-}
+.containerTop {
 
-.text {
-  font-weight: 100;
-}
-
-.h1 {
-  font-weight: 500;
 }
