+++ src/pages/auth/SignIn/SignIn.tsx
@@ -23,7 +23,7 @@ const schema = object({
 
 export const SignIn = () => {
   const navigate = useNavigate();
-  const { setIsUserLogIn, setUserEmail, setUserID } = useContext(Context);
+  const { setIsUserLogIn, setUserEmail, setUserID, setProfile } = useContext(Context);
 
   const {
     register,
@@ -47,6 +47,8 @@ export const SignIn = () => {
         setIsUserLogIn(true);
         setUserID(user.attributes.sub);
         setUserEmail(user.attributes.email);
+        setProfile(user.attributes.profile);
+
         user.attributes.profile === "doctor"
           ? navigate(`/doctor-account/${user.attributes.sub}`)
           : navigate(`/patient-account/${user.attributes.sub}`);
